Fix ESP32 demo crash, improve doxygen documentation of re-usable components (#152)
[ Description ]

(Please ignore the name of the branch; This PR has been accelerated for the v1.3 release)

The crash I had been tracking for a while was fixed by an SDK update.
Re-ordered some functions for better readability
Tiny improvements in the codebase - better return values in a couple functions, removed an unnecessary one, etc.
The documentation describes how LwIP will leave open sockets in a broken state after a sudden WiFi disconnection. In order to solve that, they suggest you close all sockets from the WiFi conn callback and re-open them on WiFi re-connection. This resulted in 2 changes to the demo:

Implemented connect/disconnect requests to the XivelyTask so we can treat WiFi first connections and reconnections in the exact same way
The Xively Task can start ticking even before the WiFi station is initialized, since it won't attempt to connect until XT_REQUEST_CONNECT is requested from the WiFi connection callback
The application now handles WiFi disconnections very gracefully.

[ JIRA ]
XCL-3072 XCL-3070 XCL-3069

[ Reviewer ]
@DellaBitta

[ QA ]
I generated the doxygen docs and took a quick look in my localhost. Everything looked fine, although the descriptions are short (and the very obvious ones are not included); that's on purpose to avoid bloating the code, so it looks 👍 to me
I've also built and run the example, disconnected and reconnected the device multiple times, left it running over night, sent lots of messages and pressed the button lots of times. It all seems to be working correctly, and the demo is extremely responsive.

[ Release Notes ]
Part of the ESP32 demo work: Write doxygen documentation for the re-usable components, update SDK, fix crashes and improve sudden WiFi disconnection handling