CBOR Assertions Now Only in Debug Mode (#169)
[ Description ]
cn-cbor contains assertions that were active in the release builds of clients. These are now only on when XI_DEBUG_ASSERT is 1 in the build environment. This is generally only toggled on during debug builds.

[ JIRA ]
None.

[ Reviewer ]
@atigyi

[ QA ]
Built and ran tests with debug and release mode.
added #error test to ensure that the debug assertion preprocessor definitions was toggled on in debug mode.

[ Release Notes ]
Assertions in CBOR encoder have been disabled in release mode.