Moved External SFT implementation from the public examples directory to the internal experimental directory. (#170)
[ Description ]
We created an External SFT implementation that could be used as a reference for writing a CBOR powered SFT implementation on 3rd party MQTT clients. Unfortunately this was being confused with our actual SFT implementation (which is internal to the client). Therefore I'm moving the external SFT implementation to the experimental directory.

[ JIRA ]
None.

[ Reviewer ]
@atigyi

[ QA ]
None.

[ Release Notes ]
External SFT client moved to the experimental directory to avoid confusion with vetted CC3200 SFT client with Xively C Client SFT/FWU BSP implementation.