Fix MQTT connection after AP reconnect in CC3220SF demo (#186)
[ Description ]

Previous behaviour:

WiFi connection OK
MQTT connection OK
[wifi AP disconnected - device's connection drops]
[wifi AP reconnected]
WiFi reconnection OK
MQTT never reconnects
By restarting the simplelink interface, we're able to reconnect to MQTT.
This is a workaround-- For a proper fix we'll have to debug the root cause, but this PR gets the device running until we have time for that.

[ JIRA ]

[ Reviewer ]
@DellaBitta

[ QA ]
Ran the application before and after the patch. The issue is resolved by this workaround.

[ Release Notes ]
Work around MQTT reconnection problems in the CC3220SF demo. Before, it would fail to reconnect to the broker after a WiFi drop