Support linking/including TLS libs outside xively-client-c (#184)
[ Description ]

Until now, libxively would download, include and configure (via build options) whichever TLS library is selected in `XI_BSP_TLS`. That was OK because we were using our own versions of mbedTLS and WolfSSL, and we could decide the build configuration.

Now, the ESP32 SDK contains a pre-built, pre-configured and hardware-accelerated  mbedTLS, so we'd like to avoid downloading/compiling/linking our own version. There are 4 steps to this implementation:

1. Create a new build option that can be set for a given PRESET/XI_BSP_TLS combination
2. Avoid downloading the TLS library when such flag is enabled. Achieved by not setting TLS_LIB_PREPARE_CMD
3. Avoid building libxively's TLS BSP with incompatible build options by removing the inclusion of `make/mt-config/mt-tls-$(XI_BSP_TLS).mk`
3. Set your custom XI_TLS_LIB_INC_DIR path to the TLS library when the flag is set.

[ JIRA ]

[ Reviewer ]
@atigyi

[ QA ]
I'm using this patch to successfully build the ESP32 PRESET using the IDF SDK's mbedTLS.
It's automatically disabled when building for XI_BSP_TLS=wolfssl, and that works fine too.

[ Release Notes ]
Support linking and inclusion of TLS libraries outside the xively-client-c directory. This is part of the effort to get the ESP32 running its SDK's version of mbedTLS, as opposed of downloading and building our own.