Feature/wolfssl implicitfallthrough (#199)
[ Description ]
Disables the implicit fall-through warnings when building wolfSSL. In the future we can re-enable them when we upgrade wolf versions (which has since added explicit fall-through instrumentation).  For now, though, the latest wolf Versions don't seem to play well with the ESP32 build, and there doesn't seem to be much need to upgrade it at this time.

[ Reviewer ]
@atigyi 

[ QA ]
Build and tested POSIX library and examples.
Built ESP32 build.

[ Release Notes ]
Build configuration change for building WolfSSL version on Ubuntu.