Bugfix/xi_desc_alloc_assert (#200)
[ Description ]
Fix in debug builds for a potential assertion scenario in our data model. The scenario would occur when the library needed to allocate a dynamically sized buffer along with the xi_data_desc_t structure to track this allocation but didn't have enough free space to allocate both.

This issue does not have a behavioral impact in release mode.

The problem would arise when the first allocation, the xi_data_desc_t, succeeds but there wasn't enough memory to subsequently create the dynamically-sized buffer. Generally both allocations take place, and then the structure of the xi_data_desc_t is initialized afterwards with the buffer size data and a pointer to the buffer. And additional data ownership field tracks who should deallocate the buffer later when the xi_data_desc_t structure is freed. This last field is important to xi_free_desc() as it needs to know if it should free the buffer that the xi_data_desc_t points to, or if someone else owns that responsibility.

In the case when the xi_data_desc_t structure was properly allocated but the subsequent dynamically allocated buffer failed in allocation, the xi_data_desc_t initialization code was skipped. When the de-allocation of the xi_data_desc_t structure occurs immediately following to cleanup the state after the error, the function xi_free_desc() would throw an assertion because the data ownership field wasn't set.

This assertion scenario has been fixed by setting up the required fields in the xi_data_desc_t before the dynamically allocated buffer malloc attempt occurs. This way if the malloc of the buffer fails, then xi_free_desc() has the information that it needs to operate on the xi_data_desc_t.

[ Reviewer ]
@atigyi

[ QA ]
Numerous Unit Tests were created to test the various data buffer creation functions in xi_data_desc.c. These new unit tests exercise the scenarios when there's not enough memory to allocate the xi_data_desc_t, and another test to if it can allocate the xi_data_desc_t tracking structure, but not the related data buffer.

[ Release Notes ]
Fixes for assertions which would occur in xi_data_desc_t creation functions when the runtime doesn't have enough free memory to allocate the buffer. This was only a debug assertion and this change has no impact on the behavior of the library in release mode.